+++ src/MultiSlider/StyledSlider.js
@@ -1,19 +1,20 @@
 import PropTypes from 'prop-types';
-import glamorous from 'glamorous';
+import styled from 'react-emotion';
 import transition from '../utils/transition';
 import getHalf from '../utils/getHalf';
 
-const StyledSlider = glamorous.div({
+const StyledSlider = styled('div')({
   position: 'relative',
   height: 14,
   boxSizing: 'border-box',
   transition
-}, ({ width, height, backgroundColor, roundedCorners, readOnly }) => ({
+}, ({ width, height, backgroundColor, roundedCorners, readOnly, css }) => ({
   width,
   height,
   backgroundColor,
   borderRadius: roundedCorners ? getHalf(height) : 0,
-  cursor: readOnly ? 'auto' : 'pointer'
+  cursor: readOnly ? 'auto' : 'pointer',
+  ...css
 }));
 
 
@@ -30,6 +31,10 @@ StyledSlider.propTypes = {
   roundedCorners: PropTypes.bool.isRequired,
   readOnly: PropTypes.bool.isRequired,
   css: PropTypes.object.isRequired
+};
+
+StyledSlider.defaultProps = {
+  css: null
 };
 
 export default StyledSlider;