+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 1.2.0
+Released 2018-09-08
+
+- Migrated from `glamorous` to `emotion`
+- Dragging now only activates when the left mouse button is clicked
+
 # Version 1.1.0
 Released 2018-03-28